requirements.txt
asgiref==3.5.2
Django==4.1.4
django-jalali==6.0.1
jdatetime==4.1.0
Pillow==9.3.0
sqlparse==0.4.3
